refactor: group diffusion stage modules
