+++ python/sglang/multimodal_gen/runtime/pipelines_core/stages/model_specific_stages/hunyuan3d/__init__.py
@@ -1,0 +1,25 @@
+# SPDX-License-Identifier: Apache-2.0
+
+"""Hunyuan3D-specific pipeline stages"""
+
+from sglang.multimodal_gen.runtime.pipelines_core.stages.model_specific_stages.hunyuan3d.paint import (
+    Hunyuan3DPaintPostprocessStage,
+    Hunyuan3DPaintPreprocessStage,
+    Hunyuan3DPaintTexGenStage,
+)
+from sglang.multimodal_gen.runtime.pipelines_core.stages.model_specific_stages.hunyuan3d.shape import (
+    Hunyuan3DShapeBeforeDenoisingStage,
+    Hunyuan3DShapeDenoisingStage,
+    Hunyuan3DShapeExportStage,
+    Hunyuan3DShapeSaveStage,
+)
+
+__all__ = [
+    "Hunyuan3DPaintPostprocessStage",
+    "Hunyuan3DPaintPreprocessStage",
+    "Hunyuan3DPaintTexGenStage",
+    "Hunyuan3DShapeBeforeDenoisingStage",
+    "Hunyuan3DShapeDenoisingStage",
+    "Hunyuan3DShapeExportStage",
+    "Hunyuan3DShapeSaveStage",
+]
