+++ python/sglang/multimodal_gen/runtime/pipelines_core/stages/realtime/__init__.py
@@ -1,0 +1,31 @@
+# SPDX-License-Identifier: Apache-2.0
+
+"""realtime pipeline stages shared by interactive diffusion models"""
+
+from sglang.multimodal_gen.runtime.pipelines_core.stages.realtime.base import (
+    RealtimeDiffusionStage,
+    RealtimeStageComponent,
+)
+from sglang.multimodal_gen.runtime.pipelines_core.stages.realtime.input_validation import (
+    RealtimeInputValidationStage,
+)
+from sglang.multimodal_gen.runtime.pipelines_core.stages.realtime.latent_preparation import (
+    RealtimeChunkLatentPreparationStage,
+)
+from sglang.multimodal_gen.runtime.pipelines_core.stages.realtime.text_encoding import (
+    RealtimeTextEncodingStage,
+)
+from sglang.multimodal_gen.runtime.pipelines_core.stages.realtime.vae import (
+    CausalVaeDecodingStage,
+    RealtimeImageVAEEncodingStage,
+)
+
+__all__ = [
+    "CausalVaeDecodingStage",
+    "RealtimeChunkLatentPreparationStage",
+    "RealtimeDiffusionStage",
+    "RealtimeImageVAEEncodingStage",
+    "RealtimeInputValidationStage",
+    "RealtimeStageComponent",
+    "RealtimeTextEncodingStage",
+]
