+++ python/sglang/multimodal_gen/runtime/pipelines_core/stages/realtime/vae.py
